Podíl SČD: podpora desetinných míst v celém projektu
Modely (Unit.podil_scd, ShareCheckRecord, SyncRecord) Integer → Float.
Parser share_check: _parse_share_value vrací float.
Excel import: podil_scd přes _cell_float.
Routery: int() → float() při ukládání z formulářů.
Nový Jinja2 filtr fmt_num — tisícové oddělovače, skrývá .0 u celých čísel.
Všechny šablony přepsány na |fmt_num.

+++ app/models/owner.py
@@ -105,7 +105,7 @@ class Unit(Base):
     id = Column(Integer, primary_key=True)
     unit_number = Column(Integer, nullable=False, unique=True, index=True)
     building_number = Column(String(20), nullable=True, index=True)
-    podil_scd = Column(Integer, nullable=True)
+    podil_scd = Column(Float, nullable=True)
     floor_area = Column(Float, nullable=True)
     room_count = Column(String(20), nullable=True)
     space_type = Column(String(50), nullable=True, index=True)
